Factored the sorting out of the saved-notebooks-and-examples-subsection.jsx component to make it dumb, and put it into getSavedNotebooks function in notebook-reducer.js. state-prototypes.js: newNotebook() function now utilizes getSavedNotebooks() function to initialize a new notebook with saved notebook state values sorted.
